Use officially exported cmake paths
PLUGIN_INSTALL_DIR is not documented in the header of KDEInstallDirs

More importantly on builds that use KDE_INSTALL_USE_QT_SYS_PATHS this is
relative, whereas normally it's absolute.

By using the CMAKE_INSTALL_FULL path we avoid that problem as it goes
through the define_relatve define_absolute macros.

REVIEW: 120603
BUG: 339539