Update most of the examples, drop outdated ones
This moves most of the example snippets (which were outdated) into real cmake targets.
We use the Doxygen @snippet magic to show them in the apidox.
This ensures that the example code is always compiled and stays up-to-date.
Some of the snippets are just removed, in particular the section about
executing the actions asynchronously, which predates the KJob (ExecuteJob) usage.

REVIEW: 129842