Remove support for passing gui QVariants to KAuth helpers
Supporting gui variants is very dangerous since they can end up triggering
image loading plugins which are one of the biggest vectors for crashes, which
for very smart people mean possible code execution, which is very dangerous
in code that is executed as root.

We've checked all the KAuth helpers inside KDE git and none seems to be using
gui variants, so we're not actually limiting anything that people wanted to do.

Reviewed by security@kde.org and Aleix Pol

Issue reported by Fabian Vogt