fix type namespace requirement
Summary:
it was hard to find in all the text anyway. and it's an entirely
pointless requirement. we can easily determine the type name dynamically
and call with or without namespace as necessary. this way users of kauth
cannot make the mistake of not using the namespace by accident.

I've thought about using a generic return argument (passing the typeName
into QGenericReturnArgument directly) but I think that opens the helper
code to memory corruption should they make the mistake of returning an
entirely unexpected type. so, what I ended up with is a bit more code
unfortunately.

Test Plan: helper code works with both `using namespace` and without

Reviewers: apol, davidedmundson, dfaure

Reviewed By: davidedmundson

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D26969