Calculate KAuthAction::isValid on demand
Summary:
Polkit1Backend fetches a list of /all/ actions once on startup, when we
want to see if an action is valid we compare it to this list.

Fine in principle except that if isValid() is called, we have to block
via a nested event loop to wait for the full list to be fetched.

Given we construct actions on startup and the backend on startup we
almost always hit this. It can easily time out (the case I'm debugging
here) or if it's invoked from QML it would simply crash.

Any good fix would be an API/behavioural break. (T12643)

In practice few places call isValid() at which point the cache is
fetched and we hide the problem, also we can document it.

It's not slower as the only backend that has
CheckActionExistenceCapability is a cache lookup anyway.

Test Plan:
Used on a slow VM with a slow KCM
My actions started working

Reviewers: broulik

Reviewed By: broulik

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27140