add helper to obtain the caller's uid
uses dbuscontext to get the remote's uid

this is handy when making an auth helper that wants to do something with
the uid. the arguments passed in aren't nearly as trustworthy as dbus in
that instance. so instead of passing in the uid, which would raise an
entire category of impersonation issues, one could get the uid from
dbus.
this way the action could also be auth_self since the call is
constrained to that uid anyway