Revert "Use QWindow instead of QWidget where possible"
Since KAuth::Action is in the "core" library, we can only use a forward
declaration of QWidget/QWindow, that means we can only use either of them in
KAuth::Action as pointers, function return types, and function parameters
but we can't use any QWidget/QWindow methods; that results in not being
able to call e.g. setParentWindow() from the setParentWidget() method, and
that made the change not backward-compatible.

Thanks to kossebau for spotting the issue.

Reverted for now, and will hopefully get applied again after KF6 branching.
Sorry for the trouble.

See the discussion in https://invent.kde.org/frameworks/kauth/-/merge_requests/9

This reverts commit 61b62617c4fbd6b502683bb266310b91790b8c8d.