Un-overload HelperProxy::progressStep() signal
By renaming the signal that takes QVariantMap arg from
HelperProxy::progressStep() to HelperProxy::progressStepData().
IIUC, this is a signal in a non-exported class.

Also rename the method HelperProxy::sendProgressStep() that takes a
QVariantMap arg to HelperProxy::sendProgressStepData().

Rationale: un-overloading the signal makes it easier to use it in connect()
calls, no need to use QOverload<>::of().

KF6 task: https://phabricator.kde.org/T13986