Add INSTALL_BROKEN_KAUTH_POLICY_FILES cmake option
It allows installing the policy files into CMAKE_INSTALL_PREFIX instead
of to kauth prefix. If those two prefixes are different, this is BAD and
will not work, but in some cases (e.g. i'm compiling all of KDE Gear and
just want to make sure things compile) I don't care, so having an option
to say "I know this is broken, but just continue" is super welcome.