authbackend: remove ExtraCallerIDVerificationMethod
with the current state of things we don't actually need it since the
only backend setting it is polkit-1 and there it is always
`VerifyAgainstDBusServiceName` which means the `callerID` MUST always be
the `message().service()` anyway.

consequently we can throw away the verification system and simply use
the service remote reported by dbus instead of the callerID argument
provided by the untrusted caller

api-wise we still need the callerID argument for the benefit of the mac
backend.