Store parent window in QPointer
Action takes an argument to a parent window. The lifespan of this is
outside the control of the Kauth action.

As this operates as a QSharedData internally it's difficult for callers
to keep track of actions and update appropriately.

This adds a QtGui dependency to KAuthCore, but we had QWindow in the
public API anyway.